Use queueMap instead of pruneMap
