+++ reconciler/reconciler.go
@@ -20,244 +20,12 @@ import (
 	"fmt"
 	"log"
 	"sort"
-	"sync"
 	"time"
 
 	"github.com/coinbase/rosetta-sdk-go/parser"
-	"github.com/coinbase/rosetta-sdk-go/storage"
 	"github.com/coinbase/rosetta-sdk-go/types"
 	"golang.org/x/sync/errgroup"
 )
-
-const (
-	// ActiveReconciliation is included in the reconciliation
-	// error message if reconciliation failed during active
-	// reconciliation.
-	ActiveReconciliation = "ACTIVE"
-
-	// InactiveReconciliation is included in the reconciliation
-	// error message if reconciliation failed during inactive
-	// reconciliation.
-	InactiveReconciliation = "INACTIVE"
-)
-
-const (
-	// BlockGone is when the block where a reconciliation
-	// is supposed to happen is orphaned.
-	BlockGone = "BLOCK_GONE"
-
-	// HeadBehind is when the synced tip (where balances
-	// were last computed) is behind the *types.BlockIdentifier
-	// returned by the call to /account/balance.
-	HeadBehind = "HEAD_BEHIND"
-
-	// BacklogFull is when the reconciliation backlog is full.
-	BacklogFull = "BACKLOG_FULL"
-
-	// TipFailure is returned when looking up the live
-	// balance fails but we are at tip. This usually occurs
-	// when the node processes an re-org that we have yet
-	// to process (so the index we are querying at may be
-	// ahead of the nodes tip).
-	TipFailure = "TIP_FAILURE"
-)
-
-const (
-	// defaultBacklogSize is the limit of account lookups
-	// that can be enqueued to reconcile before new
-	// requests are dropped.
-	defaultBacklogSize = 250000
-
-	// waitToCheckDiff is the syncing difference (live-head)
-	// to retry instead of exiting. In other words, if the
-	// processed head is behind the live head by <
-	// waitToCheckDiff we should try again after sleeping.
-	// TODO: Make configurable
-	waitToCheckDiff = 10
-
-	// waitToCheckDiffSleep is the amount of time to wait
-	// to check a balance difference if the syncer is within
-	// waitToCheckDiff from the block a balance was queried at.
-	waitToCheckDiffSleep = 5 * time.Second
-
-	// zeroString is a string of value 0.
-	zeroString = "0"
-
-	// inactiveReconciliationSleep is used as the time.Duration
-	// to sleep when there are no seen accounts to reconcile.
-	inactiveReconciliationSleep = 1 * time.Second
-
-	// defaultInactiveFrequency is the minimum
-	// number of blocks the reconciler should wait between
-	// inactive reconciliations for each account.
-	defaultInactiveFrequency = 200
-
-	// defaultReconcilerConcurrency is the number of goroutines
-	// to start for reconciliation. Half of the goroutines are assigned
-	// to inactive reconciliation and half are assigned to active
-	// reconciliation.
-	defaultReconcilerConcurrency = 8
-
-	// safeBalancePruneDepth is the depth from the last balance
-	// change that we consider safe to prune. We are very conservative
-	// here to prevent removing balances we may need in a reorg.
-	safeBalancePruneDepth = int64(500) // nolint:gomnd
-)
-
-// Helper functions are used by Reconciler to compare
-// computed balances from a block with the balance calculated
-// by the node. Defining an interface allows the client to determine
-// what sort of storage layer they want to use to provide the required
-// information.
-type Helper interface {
-	DatabaseTransaction(ctx context.Context) storage.DatabaseTransaction
-
-	CurrentBlock(
-		ctx context.Context,
-		dbTx storage.DatabaseTransaction,
-	) (*types.BlockIdentifier, error)
-
-	IndexAtTip(
-		ctx context.Context,
-		index int64,
-	) (bool, error)
-
-	CanonicalBlock(
-		ctx context.Context,
-		dbTx storage.DatabaseTransaction,
-		block *types.BlockIdentifier,
-	) (bool, error)
-
-	ComputedBalance(
-		ctx context.Context,
-		dbTx storage.DatabaseTransaction,
-		account *types.AccountIdentifier,
-		currency *types.Currency,
-		index int64,
-	) (*types.Amount, error)
-
-	LiveBalance(
-		ctx context.Context,
-		account *types.AccountIdentifier,
-		currency *types.Currency,
-		index int64,
-	) (*types.Amount, *types.BlockIdentifier, error)
-
-	// PruneBalances is invoked by the reconciler
-	// to indicate that all historical balance states
-	// <= to index can be removed.
-	PruneBalances(
-		ctx context.Context,
-		account *types.AccountIdentifier,
-		currency *types.Currency,
-		index int64,
-	) error
-}
-
-// Handler is called by Reconciler after a reconciliation
-// is performed. When a reconciliation failure is observed,
-// it is up to the client to halt syncing or log the result.
-type Handler interface {
-	ReconciliationFailed(
-		ctx context.Context,
-		reconciliationType string,
-		account *types.AccountIdentifier,
-		currency *types.Currency,
-		computedBalance string,
-		liveBalance string,
-		block *types.BlockIdentifier,
-	) error
-
-	ReconciliationSucceeded(
-		ctx context.Context,
-		reconciliationType string,
-		account *types.AccountIdentifier,
-		currency *types.Currency,
-		balance string,
-		block *types.BlockIdentifier,
-	) error
-
-	ReconciliationExempt(
-		ctx context.Context,
-		reconciliationType string,
-		account *types.AccountIdentifier,
-		currency *types.Currency,
-		computedBalance string,
-		liveBalance string,
-		block *types.BlockIdentifier,
-		exemption *types.BalanceExemption,
-	) error
-
-	ReconciliationSkipped(
-		ctx context.Context,
-		reconciliationType string,
-		account *types.AccountIdentifier,
-		currency *types.Currency,
-		cause string,
-	) error
-}
-
-// InactiveEntry is used to track the last
-// time that an *types.AccountCurrency was reconciled.
-type InactiveEntry struct {
-	Entry     *types.AccountCurrency
-	LastCheck *types.BlockIdentifier
-}
-
-// Reconciler contains all logic to reconcile balances of
-// types.AccountIdentifiers returned in types.Operations
-// by a Rosetta Server.
-type Reconciler struct {
-	helper  Helper
-	handler Handler
-	parser  *parser.Parser
-
-	lookupBalanceByBlock bool
-	interestingAccounts  []*types.AccountCurrency
-	backlogSize          int
-	changeQueue          chan *parser.BalanceChange
-	inactiveFrequency    int64
-	debugLogging         bool
-	balancePruning       bool
-
-	// Reconciler concurrency is separated between
-	// active and inactive concurrency to allow for
-	// fine-grained tuning of reconciler behavior.
-	// When there are many transactions in a block
-	// on a resource-constrained machine (laptop),
-	// it is useful to allocate more resources to
-	// active reconciliation as it is synchronous
-	// (when lookupBalanceByBlock is enabled).
-	ActiveConcurrency   int
-	InactiveConcurrency int
-
-	// highWaterMark is used to skip requests when
-	// we are very far behind the live head.
-	highWaterMark int64
-
-	// seenAccounts are stored for inactive account
-	// reconciliation. seenAccounts must be stored
-	// separately from inactiveQueue to prevent duplicate
-	// accounts from being added to the inactive reconciliation
-	// queue. If this is not done, it is possible a goroutine
-	// could be processing an account (not in the queue) when
-	// we do a lookup to determine if we should add to the queue.
-	seenAccounts  map[string]struct{}
-	inactiveQueue []*InactiveEntry
-
-	// inactiveQueueMutex needed because we can't peek at the tip
-	// of a channel to determine when it is ready to look at.
-	inactiveQueueMutex sync.Mutex
-
-	// lastIndexChecked is the last block index reconciled actively.
-	lastIndexMutex   sync.Mutex
-	lastIndexChecked int64
-
-	// pruneMap ensures we don't accidentally attempt to prune
-	// computed balances being used by other goroutines.
-	pruneMap      map[string]map[int64]int
-	pruneMapMutex sync.Mutex
-}
 
 // New creates a new Reconciler.
 func New(
@@ -278,7 +46,7 @@ func New(
 		inactiveQueue:       []*InactiveEntry{},
 		backlogSize:         defaultBacklogSize,
 		lastIndexChecked:    -1,
-		pruneMap:            map[string]map[int64]int{},
+		queueMap:            map[string]map[int64]int{},
 	}
 
 	for _, opt := range options {
@@ -338,22 +106,22 @@ func (r *Reconciler) wrappedInactiveEnqueue(
 	}
 }
 
-// addToPruneMap adds a *types.AccountCurrency
+// addToqueueMap adds a *types.AccountCurrency
 // to the prune map at the provided index.
-func (r *Reconciler) addToPruneMap(
+func (r *Reconciler) addToqueueMap(
 	acctCurrency *types.AccountCurrency,
 	index int64,
 ) {
-	r.pruneMapMutex.Lock()
+	r.queueMapMutex.Lock()
 	key := types.Hash(acctCurrency)
-	if _, ok := r.pruneMap[key]; !ok {
-		r.pruneMap[key] = map[int64]int{}
+	if _, ok := r.queueMap[key]; !ok {
+		r.queueMap[key] = map[int64]int{}
 	}
-	if _, ok := r.pruneMap[key][index]; !ok {
-		r.pruneMap[key][index] = 0
+	if _, ok := r.queueMap[key][index]; !ok {
+		r.queueMap[key][index] = 0
 	}
-	r.pruneMap[key][index]++
-	r.pruneMapMutex.Unlock()
+	r.queueMap[key][index]++
+	r.queueMapMutex.Unlock()
 }
 
 // QueueChanges enqueues a slice of *BalanceChanges
@@ -418,9 +186,9 @@ func (r *Reconciler) QueueChanges(
 			return err
 		}
 
-		// Add change to pruneMap before enqueuing to ensure
+		// Add change to queueMap before enqueuing to ensure
 		// there is no possible race.
-		r.addToPruneMap(acctCurrency, change.Block.Index)
+		r.addToqueueMap(acctCurrency, change.Block.Index)
 
 		// Add change to active queue
 		r.wrappedActiveEnqueue(ctx, change)
@@ -797,34 +565,34 @@ func (r *Reconciler) skipAndPrune(
 		return err
 	}
 
-	return r.removeAndPrune(ctx, change)
+	return r.updateQueueMapAndPrune(ctx, change)
 }
 
-// removeAndPrune removes a *parser.BalanceChange
-// from the pruneMap and attempts to prune the associated
+// updateQueueMapAndPrune removes a *parser.BalanceChange
+// from the queueMap and attempts to prune the associated
 // *types.AccountCurrency's balances, if appropriate.
-func (r *Reconciler) removeAndPrune(
+func (r *Reconciler) updateQueueMapAndPrune(
 	ctx context.Context,
 	change *parser.BalanceChange,
 ) error {
-	r.pruneMapMutex.Lock()
+	r.queueMapMutex.Lock()
 
 	key := types.Hash(&types.AccountCurrency{
 		Account:  change.Account,
 		Currency: change.Currency,
 	})
-	r.pruneMap[key][change.Block.Index]--
-	if r.pruneMap[key][change.Block.Index] > 0 {
-		r.pruneMapMutex.Unlock()
+	r.queueMap[key][change.Block.Index]--
+	if r.queueMap[key][change.Block.Index] > 0 {
+		r.queueMapMutex.Unlock()
 		return nil
 	}
 
 	// Cleanup indexes when we don't need them anymore
-	delete(r.pruneMap[key], change.Block.Index)
+	delete(r.queueMap[key], change.Block.Index)
 
 	// Sort indexes
 	indexes := []int64{}
-	for k := range r.pruneMap[key] {
+	for k := range r.queueMap[key] {
 		indexes = append(indexes, k)
 	}
 	sort.Slice(indexes, func(i, j int) bool { return indexes[i] < indexes[j] })
@@ -832,17 +600,17 @@ func (r *Reconciler) removeAndPrune(
 	// Don't prune if there are indexes for this AccountCurrency
 	// less than this change.
 	if len(indexes) > 0 && change.Block.Index >= indexes[0] {
-		r.pruneMapMutex.Unlock()
+		r.queueMapMutex.Unlock()
 		return nil
 	}
 
 	// Cleanup keys when we don't need them anymore
-	if len(r.pruneMap[key]) == 0 {
-		delete(r.pruneMap, key)
+	if len(r.queueMap[key]) == 0 {
+		delete(r.queueMap, key)
 	}
 
 	// Unlock before pruning as this could take some time
-	r.pruneMapMutex.Unlock()
+	r.queueMapMutex.Unlock()
 
 	// Attempt to prune historical balances that will not be used
 	// anymore.
@@ -921,7 +689,7 @@ func (r *Reconciler) reconcileActiveAccounts(ctx context.Context) error { // nol
 
 			// Attempt to prune historical balances that will not be used
 			// anymore.
-			if err := r.removeAndPrune(ctx, balanceChange); err != nil {
+			if err := r.updateQueueMapAndPrune(ctx, balanceChange); err != nil {
 				return err
 			}
